+++ txmongo/collection.py
@@ -597,6 +597,11 @@ class Collection:
 
     @timeout
     def group(self, keys, initial, reduce, condition=None, finalize=None, **kwargs):
+        warnings.warn(
+            "Collection.group() method will be removed in the next version of TxMongo. Please use aggregate() or map_reduce().",
+            DeprecationWarning,
+        )
+
         body = {
             "ns": self._collection_name,
             "initial": initial,
@@ -1086,6 +1091,12 @@ class Collection:
 
     @timeout
     def save(self, doc, safe=None, **kwargs):
+        warnings.warn(
+            "Collection.save() method will be removed in the next version of TxMongo. "
+            "Please use insert_one() or replace_one().",
+            DeprecationWarning,
+        )
+
         if not isinstance(doc, dict):
             raise TypeError(
                 "TxMongo: cannot save objects of type {0}".format(type(doc))
@@ -1338,6 +1349,12 @@ class Collection:
 
     @timeout
     def find_and_modify(self, query=None, update=None, upsert=False, **kwargs):
+        warnings.warn(
+            "Collection.find_and_modify() method will be removed in the next version of TxMongo. "
+            "Please use find_one_and_update(), find_one_and_replace() or find_one_and_delete().",
+            DeprecationWarning,
+        )
+
         no_obj_error = "No matching object found"
 
         if not update and not kwargs.get("remove", None):